requirements.txt
pyxnat==1.6
cryptography==41.0.3
pathlib==1.0.1
cffi==1.15.1
future==0.18.3
Pillow==10.0.0
pip==21.3.1
configparser==6.0.0
wheel==0.37.1
lxml==4.9.3
rsa==4.9
pyasn1==0.5.0
six==1.16.0
prov==2.0.0
rdflib==7.0.0
pydot==1.4.2
networkx==3.1
python-dateutil==2.8.2
pytz==2023.3
idna==3.4
multidict==6.0.4
attrs==23.1.0
MarkupSafe==2.1.3
Werkzeug==2.3.6
click==8.1.6
blinker==1.6.2
Jinja2==3.1.2
itsdangerous==2.1.2
numpy==1.24.4
setuptools==60.2.0
panda==0.3.1
requests==2.31.0
scipy==1.10.1
github==1.2.7
PyGithub==1.59.1
PyJWT==2.8.0
aiohttp==3.8.1
Deprecated==1.2.14
urllib3==1.26.16
cachetools==5.3.1
certifi==2023.7.22
oauth2client==4.1.3
nipype==1.8.6
nibabel==5.1.0
looseversion==1.3.0
simplejson==3.19.1
filelock==3.12.2
pandas==2.0.3
traits==6.3.2
packaging==23.1
openpyxl==3.1.2
pyparsing==3.1.0
isodate==0.6.1
yarl==1.9.2
aiosignal==1.3.1
frozenlist==1.4.0
httplib2==0.22.0
oauthlib==3.2.2
configobj==5.0.8
cairocffi==1.6.1
pycparser==2.21
wrapt==1.15.0
Flask==2.3.2
zipp==3.16.2
gspread==5.10.0